restructure link component
