+++ src/app/page.tsx
@@ -4,7 +4,7 @@ import clsx from 'clsx';
 import Image from 'next/image';
 import { useState } from 'react';
 
-import Link from '@/component/Link';
+import Link from '@/component/primitive/Link';
 import SignupModal from '@/component/SignupModal';
 
 export default function Home() {
